[doc] Include samples in doxygen, reorganize and install samples
