+++ src/mc_control/CMakeLists.txt
@@ -52,6 +52,7 @@ if(${BUILD_CONTROLLER_SAMPLES})
   add_subdirectory(samples/Admittance)
   add_subdirectory(samples/Impedance)
   add_subdirectory(samples/ExternalForces)
+  add_subdirectory(samples/CoMTrajectoryGeneration)
 endif()
 
 add_subdirectory(HalfSitPose)
