+++ src/mc_control/samples/CoMTrajectoryGeneration/src/CMakeLists.txt
@@ -8,7 +8,7 @@ set(controller_HDR
 
 set(CONTROLLER_NAME CoMTrajectoryGeneration)
 add_library(${CONTROLLER_NAME} SHARED ${controller_SRC} ${controller_HDR})
-set_target_properties(${CONTROLLER_NAME} PROPERTIES COMPILE_FLAGS "-DCoMTrajectoryGeneration_EXPORTS")
+set_target_properties(${CONTROLLER_NAME} PROPERTIES COMPILE_FLAGS "-DMC_CONTROL_EXPORTS")
 target_link_libraries(${CONTROLLER_NAME} PUBLIC mc_rtc::mc_control_fsm)
 install(TARGETS ${CONTROLLER_NAME} DESTINATION ${MC_CONTROLLER_INSTALL_PREFIX}/..)
