+++ src/mc_control/samples/CoMTrajectoryGeneration/src/states/StabilizerStandingTrackCoM.h
@@ -16,17 +16,17 @@ namespace mc_samples
 struct StabilizerStandingTrackCoM : mc_control::fsm::State
 {
 
-    void configure(const mc_rtc::Configuration & config) override;
+  void configure(const mc_rtc::Configuration & config) override;
 
-    void start(mc_control::fsm::Controller & ctl) override;
+  void start(mc_control::fsm::Controller & ctl) override;
 
-    bool run(mc_control::fsm::Controller & ctl) override;
+  bool run(mc_control::fsm::Controller & ctl) override;
 
-    void teardown(mc_control::fsm::Controller & ctl) override;
+  void teardown(mc_control::fsm::Controller & ctl) override;
 
 private:
   std::shared_ptr<mc_tasks::lipm_stabilizer::StabilizerTask> stabilizerTask_ = nullptr;
   mc_rtc::Configuration config_;
 };
 
-} /* mc_samples */
+} // namespace mc_samples
