--- tests/hmc/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_executable(generator testCoMGenerator.cpp)
-target_link_libraries(generator PUBLIC mc_planning)
